refactor: changes app structure, refactor code
