+++ src/album/album.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@nestjs/common';
 import { v4 as uuid4 } from 'uuid';
 import { CreateAlbumDto, UpdateAlbumDto } from './album.dto';
-import { Album } from '../types/types';
-import { DBService } from '../db/db';
+import { Album } from '../types';
+import { DBService } from '../db';
 
 @Injectable()
 export class AlbumService {
