+++ src/app.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { DBModule } from './db/db';
+import { DBModule } from './db';
 import { UserModule } from './user/user.modul';
 import { ArtistModule } from './artist/artist.modul';
 import { AlbumModule } from './album/album.modul';
