+++ src/types.ts
@@ -34,8 +34,6 @@ export interface Favorites {
   tracks: string[];
 }
 
-export type UserResp = Omit<User, 'password'>;
-
 export enum MessageStatus {
   USER_NOT_FOUND = 'User not found',
   ARTIST_NOT_FOUND = 'Artist not found',
@@ -43,8 +41,9 @@ export enum MessageStatus {
   TRACK_NOT_FOUND = 'Track not found',
   USER_PASSWORD_IS_INVALID = 'User password is invalid',
 }
+export type UserResp = Omit<User, 'password'>;
 
-export interface FavoritesRepsonse {
+export interface FavoritesResp {
   artists: Artist[];
   albums: Album[];
   tracks: Track[];
